require permission to delete uneeded files
changed author

+++ main.py
@@ -122,9 +122,10 @@ def main() -> None:
             print(f"  □ {item}")
         print("="*50 + "\n")
     
-    logger.info("Cleaning up uneeded files...")
-    os.remove(video_path)
-    os.remove(audio_path)
+    if bool(os.environ("DELETE_UNEEDED_FILES")):
+        logger.info("Cleaning up uneeded files...")
+        os.remove(video_path)
+        os.remove(audio_path)
     logger.info("Successfully ran program...")
